Final Fixes 1.0
Fixed case-sensitivity, being able to DM the bot "test" to crash it, multiple people being able to trigger the bot while it was making the video, and not being able to invite the bot to servers (on Discord's end).

Added typing... then the bot is making a new video and a better readme.

Changed the default word from 'test' to 'klaxon'.